Refines institutional data schemas and metrics
Standardizes column names related to geography, population, and area across institutional presets.
Introduces new socio-economic metrics (PMT, NBI) and total intervention counts.
Updates institutional data sources to use a dedicated view and more precise base filters, specifically for MIDES.
Aligns allowed filter fields with the updated column naming conventions.
Adjusts the boolean type mapping for database compatibility.
Ensures all detail columns are correctly included during data source seeding.

+++ api/v1/services/query_engine/engine.py
@@ -4,7 +4,7 @@ _CH_TYPE_MAP = {
     "TEXT": "String",
     "INTEGER": "Int64",
     "FLOAT": "Float64",
-    "BOOLEAN": "Int8",
+    "BOOLEAN": "Int32",
 }
 
 _OP_MAP = {
